adding some logs to the system

+++ src/main/java/com/finalproject/tuwaiqfinal/Service/HallService.java
@@ -1,15 +1,12 @@
 package com.finalproject.tuwaiqfinal.Service;
 
 import com.finalproject.tuwaiqfinal.Api.ApiException;
-import com.finalproject.tuwaiqfinal.DTOin.OwnerDTO;
 import com.finalproject.tuwaiqfinal.Model.Hall;
 import com.finalproject.tuwaiqfinal.Model.Owner;
 import com.finalproject.tuwaiqfinal.Model.SubHall;
-import com.finalproject.tuwaiqfinal.Model.User;
 import com.finalproject.tuwaiqfinal.Repository.HallRepository;
 import com.finalproject.tuwaiqfinal.Repository.OwnerRepository;
 import com.finalproject.tuwaiqfinal.Repository.SubHallRepository;
-import com.finalproject.tuwaiqfinal.Repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -56,13 +53,23 @@ public class HallService {
         Owner owner = ownerRepository.findById(ownerId)
                 .orElseThrow(()-> new ApiException("owner not found"));
 
+        if (!hall.getOwner().equals(owner))
+            throw new ApiException("permission denied");
+
         oldHall.setOwner(owner);
         hallRepository.save(oldHall);
     }
 
-    public void deleteHall(Integer hallId){
+    public void deleteHall(Integer ownerId,Integer hallId){
         Hall hall = hallRepository.findById(hallId)
                 .orElseThrow(()-> new ApiException("hall not found"));
+
+        Owner owner = ownerRepository.findById(ownerId)
+                        .orElseThrow(()-> new ApiException("owner not found"));
+
+        if (!hall.getOwner().equals(owner))
+            throw new ApiException("permission denied");
+
         hall.setOwner(null);
         hallRepository.delete(hall);
     }
